+++ src/main/kotlin/org/cubewhy/celestial/entity/vo/AnalysisVO.kt
@@ -6,6 +6,6 @@ import java.time.Instant
 data class AnalysisVO(
     var userCount: Long,
     var webUserCount: Long,
-    var onlineCount: Int,
+    var onlineCount: Long,
     val timestamp: Instant
 ): Serializable
